CPE: correct field types for IMAGE_DOS_HEADER

+++ Sources/CPE/include/CPE.h
@@ -14,24 +14,24 @@
 #define IMAGE_SIZEOF_SHORT_NAME 8
 
 typedef struct _IMAGE_DOS_HEADER {
-   uint8_t  e_magic;
-   uint8_t  e_cblp;
-   uint8_t  e_cp;
-   uint8_t  e_crlc;
-   uint8_t  e_cparhdr;
-   uint8_t  e_minalloc;
-   uint8_t  e_maxalloc;
-   uint8_t  e_ss;
-   uint8_t  e_sp;
-   uint8_t  e_csum;
-   uint8_t  e_ip;
-   uint8_t  e_cs;
-   uint8_t  e_lfarlc;
-   uint8_t  e_ovno;
-   uint8_t  e_res[4];
-   uint8_t  e_oemid;
-   uint8_t  e_oeminfo;
-   uint8_t  e_res2[10];
+  uint16_t  e_magic;
+  uint16_t  e_cblp;
+  uint16_t  e_cp;
+  uint16_t  e_crlc;
+  uint16_t  e_cparhdr;
+  uint16_t  e_minalloc;
+  uint16_t  e_maxalloc;
+  uint16_t  e_ss;
+  uint16_t  e_sp;
+  uint16_t  e_csum;
+  uint16_t  e_ip;
+  uint16_t  e_cs;
+  uint16_t  e_lfarlc;
+  uint16_t  e_ovno;
+  uint16_t  e_res[4];
+  uint16_t  e_oemid;
+  uint16_t  e_oeminfo;
+  uint16_t  e_res2[10];
   uint32_t  e_lfanew;
 } IMAGE_DOS_HEADER, *PIMAGE_DOS_HEADER;
